requirements.txt
pandas==1.5.3
rapidfuzz==2.8.0
python-dateutil==2.8.2
numpy==1.24.0
